Fix critical race conditions and optimize allocations
Thread Safety Fixes:
- Fix race conditions in GetOrCreateArchetype and GetMatchingArchetypes by always locking
- Fix GroupResultEnumerator to get fresh entity span on each iteration for concurrent modifications
- Add volatile fields to Singleton<T> for cross-thread visibility
- Add global lock and volatile fields to TagRegistry for thread-safe static registration

Component Index Upgrade:
- Change component index from byte to ushort (supports up to 65,536 component types)
- Update ComponentRef, TagRegistry, and all related code to use ushort

Allocation Elimination:
- Remove string interpolation from exception messages in hot paths
- Remove typeof(T) calls from error messages to prevent boxing allocations
- Simplify exception messages to avoid dynamic string construction

All changes maintain zero-allocation design and thread-safe immediate execution.

+++ EasyEcs.Core/Components/ComponentRef.cs
@@ -12,10 +12,10 @@ public readonly struct ComponentRef<T> where T : struct, IComponent
 {
     // Pack entity ID (lower 32 bits) + version (upper 32 bits)
     private readonly long _packed;
-    private readonly byte _componentIndex;
+    private readonly ushort _componentIndex;  // Supports up to 65536 component types
     private readonly Context _context;
 
-    public ComponentRef(int entityId, int version, byte componentIndex, Context context)
+    public ComponentRef(int entityId, int version, ushort componentIndex, Context context)
     {
         _packed = ((long)version << 32) | (uint)entityId;
         _componentIndex = componentIndex;
